update rn to 37 on example
